LEN may occur as an INTRINSIC_OP or alternatively as a CALL. It seems that the mfef90-generated LEN references (for the internal passing of string length arguments, for example) are usually INTRINSIC_OPs while LEN calls explcitly occuring in the input code are usually CALLs. Becaus of that we had to introduce a key and needed to add attaching the key as an annotation for the INTRINISIC_OP case. We also introduced LEN.
For the substring arguments we had to handle the case where
we call
LEN(TRIM(aString))
and the intermediate treetype did not evaluate to a pointer.
For the same case we also had to add the condition to the
VARREF translation to avoid finding the nested TRIM call
ending up within a VarRef context.